Split josh-proxy service code
This gives us ability to create josh-proxy HTTP server instance, along
with all the associated data, from e.g. benchmarks, by calling
make_service_router. The created instance will not have all the extra
features like background polling for example, so it will be convenient
for those automated test setups.
