+++ josh-proxy/src/lib.rs
@@ -5,13 +5,17 @@ pub mod housekeeping;
 pub mod http;
 pub mod service;
 pub mod trace;
+pub mod upstream;
 
-use josh_core::cache::{CacheStack, TransactionContext};
-use josh_core::{JoshError, JoshResult, josh_error};
-use std::fs;
-use std::path::PathBuf;
+use josh_core::{JoshError, josh_error};
 
-use josh_core::changes::PushMode;
+use crate::upstream::RemoteAuth;
+
+josh_core::regex_parsed!(
+    FilteredRepoUrl,
+    r"(?P<api>/~/\w+)?(?P<upstream_repo>/[^:!]*[.]git)(?P<headref>[\^@][^:!]*)?((?P<filter_spec>[:!].*)[.]git)?(?P<pathinfo>/.*)?(?P<rest>.*)",
+    [api, upstream_repo, filter_spec, pathinfo, headref, rest]
+);
 
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, Default)]
 pub struct Ref {
@@ -50,480 +54,6 @@ pub fn refs_locking(refs: Vec<(String, git2::Oid)>, meta: &MetaConfig) -> Vec<(S
     }
 
     output
-}
-
-use josh_core::changes::baseref_and_options;
-
-#[derive(serde::Serialize, serde::Deserialize, Debug, Clone)]
-pub enum RemoteAuth {
-    Http { auth: auth::Handle },
-    Ssh { auth_socket: PathBuf },
-}
-
-#[derive(serde::Serialize, serde::Deserialize, Debug)]
-pub struct RepoUpdate {
-    pub refs: std::collections::HashMap<String, (String, String)>,
-    pub remote_url: String,
-    pub remote_auth: RemoteAuth,
-    pub port: String,
-    pub filter_spec: String,
-    pub base_ns: String,
-    pub git_ns: String,
-    pub git_dir: String,
-    pub mirror_git_dir: String,
-    pub context_propagator: std::collections::HashMap<String, String>,
-}
-
-#[derive(serde::Serialize, serde::Deserialize, Debug)]
-#[serde(default)]
-#[derive(Default)]
-pub struct PushOptions {
-    pub merge: bool,
-    pub allow_orphans: bool,
-    pub edit: bool,
-    pub create: bool,
-    pub force: bool,
-    pub base: Option<String>,
-    pub author: Option<String>,
-}
-
-pub fn process_repo_update(repo_update: RepoUpdate) -> josh_core::JoshResult<String> {
-    let push_options_path = std::path::PathBuf::from(&repo_update.git_dir)
-        .join("refs/namespaces")
-        .join(&repo_update.git_ns)
-        .join("push_options");
-
-    let push_options = std::fs::read_to_string(&push_options_path)?;
-    std::fs::remove_file(push_options_path).ok();
-
-    let push_options: PushOptions = serde_json::from_str(&push_options)
-        .map_err(|e| josh_error(&format!("Failed to parse push options: {}", e)))?;
-
-    tracing::debug!(
-        push_options = ?push_options,
-        "process_repo_update"
-    );
-
-    let cache = std::sync::Arc::new(CacheStack::default());
-    let transaction_ctx = TransactionContext::new(&repo_update.git_dir, cache.clone());
-    let transaction_mirror_ctx =
-        TransactionContext::new(&repo_update.mirror_git_dir, cache.clone());
-
-    for (refname, (old, new)) in repo_update.refs.iter() {
-        let transaction = transaction_ctx.open(Some(&format!(
-            "refs/josh/upstream/{}/",
-            repo_update.base_ns
-        )))?;
-
-        let transaction_mirror = transaction_mirror_ctx.open(Some(&format!(
-            "refs/josh/upstream/{}/",
-            repo_update.base_ns
-        )))?;
-
-        transaction.repo().odb()?.add_disk_alternate(
-            transaction_mirror
-                .repo()
-                .path()
-                .join("objects")
-                .to_str()
-                .unwrap(),
-        )?;
-
-        let old = git2::Oid::from_str(old)?;
-        let (baseref, push_to, options, push_mode) = baseref_and_options(refname)?;
-
-        let old = if old == git2::Oid::zero() {
-            let rev = format!("refs/namespaces/{}/{}", repo_update.git_ns, &baseref);
-            let oid = if let Ok(resolved) = transaction.repo().revparse_single(&rev) {
-                resolved.id()
-            } else {
-                old
-            };
-
-            tracing::debug!(
-                old_oid = ?oid,
-                rev = %rev,
-                "resolve_old"
-            );
-
-            oid
-        } else {
-            tracing::debug!(
-                old_oid = ?old,
-                refname = %refname,
-                "resolve_old"
-            );
-
-            old
-        };
-
-        let original_target_ref = if let Some(base) = &push_options.base {
-            // Allow user to use just the branch name as the base:
-            let full_path_base_refname =
-                transaction_mirror.refname(&format!("refs/heads/{}", base));
-            if transaction_mirror
-                .repo()
-                .refname_to_id(&full_path_base_refname)
-                .is_ok()
-            {
-                full_path_base_refname
-            } else {
-                transaction_mirror.refname(base)
-            }
-        } else {
-            transaction_mirror.refname(&baseref)
-        };
-
-        let original_target = if let Ok(oid) = transaction_mirror
-            .repo()
-            .refname_to_id(&original_target_ref)
-        {
-            tracing::debug!(
-                original_target_oid = ?oid,
-                original_target_ref = %original_target_ref,
-                "resolve_original_target"
-            );
-
-            oid
-        } else {
-            tracing::debug!(
-                original_target_ref = %original_target_ref,
-                "resolve_original_target"
-            );
-
-            return Err(josh_core::josh_error(&indoc::formatdoc!(
-                r###"
-                    Reference {:?} does not exist on remote.
-                    If you want to create it, pass "-o base=<basebranch>" or "-o base=path/to/ref"
-                    to specify a base branch/reference.
-                    "###,
-                baseref
-            )));
-        };
-
-        let reparent_orphans = if push_options.create {
-            Some(original_target)
-        } else {
-            None
-        };
-
-        let author = if let Some(author) = &push_options.author {
-            author.to_string()
-        } else {
-            "".to_string()
-        };
-
-        let mut changes =
-            if push_mode == PushMode::Stack || push_mode == PushMode::Split || !author.is_empty() {
-                Some(vec![])
-            } else {
-                None
-            };
-
-        let filter = josh_core::filter::parse(&repo_update.filter_spec)?;
-        let new_oid = git2::Oid::from_str(new)?;
-        let backward_new_oid = {
-            let unapply_result = josh_core::history::unapply_filter(
-                &transaction,
-                filter,
-                original_target,
-                old,
-                new_oid,
-                if push_options.merge || push_options.allow_orphans {
-                    josh_core::history::OrphansMode::Keep
-                } else if push_options.edit {
-                    josh_core::history::OrphansMode::Remove
-                } else {
-                    josh_core::history::OrphansMode::Fail
-                },
-                reparent_orphans,
-                &mut changes,
-            )?;
-
-            tracing::debug!(
-                processed_old = ?old,
-                unapply_result = ?unapply_result,
-                "unapply_filter"
-            );
-
-            unapply_result
-        };
-
-        let oid_to_push = if push_options.merge {
-            let backward_commit = transaction.repo().find_commit(backward_new_oid)?;
-            if let Ok(base_commit_id) = transaction_mirror
-                .repo()
-                .revparse_single(&original_target_ref)
-                .map(|x| x.id())
-            {
-                let signature = proxy_commit_signature()?;
-                let base_commit = transaction.repo().find_commit(base_commit_id)?;
-                let merged_tree = transaction
-                    .repo()
-                    .merge_commits(&base_commit, &backward_commit, None)?
-                    .write_tree_to(transaction.repo())?;
-                transaction.repo().commit(
-                    None,
-                    &signature,
-                    &signature,
-                    &format!("Merge from {}", &repo_update.filter_spec),
-                    &transaction.repo().find_tree(merged_tree)?,
-                    &[&base_commit, &backward_commit],
-                )?
-            } else {
-                return Err(josh_core::josh_error("josh_merge failed"));
-            }
-        } else {
-            backward_new_oid
-        };
-
-        let ref_with_options = if !options.is_empty() {
-            format!("{}{}{}", push_to, "%", options.join(","))
-        } else {
-            push_to
-        };
-
-        let to_push = build_to_push(
-            transaction.repo(),
-            changes,
-            push_mode,
-            &baseref,
-            &author,
-            &ref_with_options,
-            oid_to_push,
-            old,
-        )?;
-
-        let mut resp = vec![];
-
-        for (reference, oid, display_name) in to_push {
-            let force_push = push_mode != PushMode::Normal || push_options.force;
-
-            let (text, status) = push_head_url(
-                transaction.repo(),
-                &format!("{}/objects", repo_update.mirror_git_dir),
-                oid,
-                &reference,
-                &repo_update.remote_url,
-                &repo_update.remote_auth,
-                &repo_update.git_ns,
-                &display_name,
-                force_push,
-            )?;
-
-            if status != 0 {
-                return Err(josh_core::josh_error(&text));
-            }
-
-            resp.push(text.to_string());
-            let mut warnings = josh_core::filter::compute_warnings(
-                &transaction,
-                filter,
-                transaction.repo().find_commit(oid)?.tree()?,
-            );
-
-            if !warnings.is_empty() {
-                resp.push("warnings:".to_string());
-                resp.append(&mut warnings);
-            }
-        }
-
-        let reapply = josh_core::filter::apply_to_commit(
-            filter,
-            &transaction.repo().find_commit(oid_to_push)?,
-            &transaction,
-        )?;
-
-        if new_oid != reapply {
-            if std::env::var("JOSH_REWRITE_REFS").is_ok() {
-                transaction.repo().reference(
-                    &format!(
-                        "refs/josh/rewrites/{}/{:?}/r_{}",
-                        repo_update.base_ns,
-                        filter.id(),
-                        reapply
-                    ),
-                    reapply,
-                    true,
-                    "reapply",
-                )?;
-            }
-
-            tracing::debug!(
-                new_oid = ?new_oid,
-                reapply = ?reapply,
-                "rewrite"
-            );
-
-            let text = format!("REWRITE({} -> {})", new_oid, reapply);
-            resp.push(text);
-        }
-
-        return Ok(resp.join("\n"));
-    }
-
-    Ok("".to_string())
-}
-
-use josh_core::changes::build_to_push;
-
-pub fn push_head_url(
-    repo: &git2::Repository,
-    alternate: &str,
-    oid: git2::Oid,
-    refname: &str,
-    url: &str,
-    remote_auth: &RemoteAuth,
-    namespace: &str,
-    display_name: &str,
-    force: bool,
-) -> josh_core::JoshResult<(String, i32)> {
-    let push_temp_ref = format!("refs/{}", &namespace);
-    let push_refspec = format!("{}:{}", &push_temp_ref, &refname);
-
-    let mut cmd = vec!["git", "push"];
-    if force {
-        cmd.push("--force")
-    }
-    cmd.push(url);
-    cmd.push(&push_refspec);
-
-    let mut fake_head = repo.reference(&push_temp_ref, oid, true, "push_head_url")?;
-    let (stdout, stderr, status) =
-        run_git_with_auth(repo.path(), &cmd, remote_auth, Some(alternate.to_owned()))?;
-    fake_head.delete()?;
-
-    tracing::debug!(
-        stdout = %stdout,
-        stderr = %stderr,
-        status = %status,
-        "push_head_url: run_git"
-    );
-
-    let stderr = stderr.replace(&push_temp_ref, display_name);
-    Ok((stderr, status))
-}
-
-fn create_repo_base(path: &PathBuf) -> josh_core::JoshResult<josh_core::shell::Shell> {
-    std::fs::create_dir_all(path).expect("can't create_dir_all");
-
-    if gix::open(path).is_err() {
-        gix::init_bare(path)?;
-    }
-
-    let credential_helper =
-        r#"!f() { echo username="${GIT_USER}"; echo password="${GIT_PASSWORD}"; }; f"#;
-
-    let config_options = [
-        ("http", &[("receivepack", "true")] as &[(&str, &str)]),
-        (
-            "user",
-            &[("name", "JOSH"), ("email", "josh@josh-project.dev")],
-        ),
-        (
-            "uploadpack",
-            &[
-                ("allowAnySHA1InWant", "true"),
-                ("allowReachableSHA1InWant", "true"),
-                ("allowTipSha1InWant", "true"),
-            ],
-        ),
-        ("receive", &[("advertisePushOptions", "true")]),
-        ("gc", &[("auto", "0")]),
-        ("credential", &[("helper", credential_helper)]),
-    ];
-
-    let shell = josh_core::shell::Shell {
-        cwd: path.to_path_buf(),
-    };
-
-    let config_source = gix::config::Source::Local;
-    let config_location = config_source.storage_location(&mut |_| None).unwrap();
-    let config_location = path.join(config_location);
-
-    let mut config =
-        gix::config::File::from_path_no_includes(config_location.clone(), config_source)
-            .map_err(|_| josh_error("unable to open repo config file"))?;
-
-    config_options
-        .iter()
-        .cloned()
-        .try_for_each(|(section, values)| -> JoshResult<()> {
-            let mut section = config
-                .new_section(section, None)
-                .map_err(|_| josh_error("unable to create config section"))?;
-
-            values
-                .iter()
-                .cloned()
-                .try_for_each(|(name, value)| -> JoshResult<()> {
-                    use gix::config::parse::section::ValueName;
-
-                    let key = ValueName::try_from(name)
-                        .map_err(|_| josh_error("unable to create config section"))?;
-                    let value = Some(value.into());
-
-                    section.push(key, value);
-
-                    Ok(())
-                })?;
-
-            Ok(())
-        })?;
-
-    fs::write(&config_location, config.to_string())?;
-
-    let hooks = path.join("hooks");
-    let packed_refs = path.join("packed-refs");
-
-    if hooks.exists() {
-        fs::remove_dir_all(hooks)?;
-    }
-
-    if packed_refs.exists() {
-        fs::remove_file(packed_refs)?;
-    }
-
-    // Delete all files ending with ".lock"
-    fs::read_dir(path)?
-        .filter_map(|entry| match entry {
-            Ok(entry) if entry.path().ends_with(".lock") => Some(path),
-            _ => None,
-        })
-        .map(fs::remove_file)
-        .collect::<Result<Vec<_>, _>>()?;
-
-    Ok(shell)
-}
-
-pub fn create_repo(path: &std::path::Path) -> josh_core::JoshResult<()> {
-    let mirror_path = path.join("mirror");
-    tracing::debug!("init mirror repo: {:?}", mirror_path);
-    create_repo_base(&mirror_path)?;
-
-    let overlay_path = path.join("overlay");
-    tracing::debug!("init overlay repo: {:?}", overlay_path);
-    let overlay_shell = create_repo_base(&overlay_path)?;
-    overlay_shell.command(&["mkdir", "hooks"]);
-
-    let josh_executable = std::env::current_exe().expect("can't find path to exe");
-    std::os::unix::fs::symlink(
-        josh_executable.clone(),
-        overlay_path.join("hooks").join("update"),
-    )
-    .expect("can't symlink update hook");
-    std::os::unix::fs::symlink(
-        josh_executable,
-        overlay_path.join("hooks").join("pre-receive"),
-    )
-    .expect("can't symlink pre-receive hook");
-
-    if std::env::var_os("JOSH_KEEP_NS").is_none() {
-        std::fs::remove_dir_all(overlay_path.join("refs/namespaces")).ok();
-    }
-
-    tracing::info!("repo initialized");
-    Ok(())
 }
 
 fn make_ssh_command() -> String {
@@ -746,7 +276,7 @@ impl Drop for TmpGitNamespace {
         let request_tmp_namespace = self.repo_path.join("refs/namespaces").join(&self.name);
 
         if std::path::Path::new(&request_tmp_namespace).exists() {
-            fs::remove_dir_all(&request_tmp_namespace).unwrap_or_else(|err| {
+            std::fs::remove_dir_all(&request_tmp_namespace).unwrap_or_else(|err| {
                 tracing::error!(
                     "remove_dir_all {} failed, error: {}",
                     request_tmp_namespace.display(),
